Write a program that multiplies two numbers

+++ 0x0A-argc_argv/main.h
@@ -1,10 +1,12 @@
 #ifndef MAIN_H
 #define MAIN_H
 
+#include <stddef.h>
+
 int _putchar(char str);
 void print_program_name(char *program_name);
 void print_program_number(int num);
 void print_args(int argc, char *argv[]);
-int mul(int n, int m);
+unsigned int mul(size_t n, size_t m);
 
 #endif
